backend/python/requirements.txt
fastapi
uvicorn
aiofile
pydantic
PyJWT